oppdatert filstruktur
satt inn i riktig mappe

gjør mellom-commit før destruktiv handlinger
