+++ careconnect/src/main/java/LoginTest/ReadMe
@@ -1,0 +1,2 @@
+Authenticate users from client to backend.
+
