+++ careconnect/src/main/java/TilstandsMonitor/Tilstandsmonitor$Grenser.class
